Renaming model to unet
